backend/requirements.txt
flake8==3.7.9
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
mypy==0.770
psycopg2==2.8.5
pytest-flask==1.0.0
requests==2.23.0
shortuuid==1.0.1
SQLAlchemy==1.3.15
flask_cors
werkzeug
PyJWT
black
